fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,13 +7,13 @@
  */
 int main(void)
 {
-	int max = 50, num1 = 1, num2 = 2, num3 = 0, i = 0;
+	long int max = 50, num1 = 1, num2 = 2, num3 = 0, i = 0;
 
-	printf("%d", num1);
+	printf("%ld", num1);
 	while (i < max)
 	{
 		num3 = num1 + num2;
-		printf(", %d", num3);
+		printf(", %ld", num3);
 		num1 = num2;
 		num2 = num3;
